refactor: Event로 publish delete 로직 분리
