+++ backend/src/main/java/hanglog/trip/domain/PublishedDeleteEvent.java
@@ -1,11 +1,11 @@
-package hanglog.community.domain;
+package hanglog.trip.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class PublishedEvent {
+public class PublishedDeleteEvent {
 
     private final Long tripId;
 }
