+++ backend/src/main/java/hanglog/trip/domain/PublishedEvent.java
@@ -1,0 +1,11 @@
+package hanglog.trip.domain;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class PublishedEvent {
+
+    private final Long tripId;
+}
